データをfloatに変換せずにLSB,MSBで保存する構造体を追加

+++ Projects/STM32duino/SensingModule/Thermistor.cpp
@@ -11,9 +11,9 @@ void Thermistor::initialize() {
 }
 
 
-void Thermistor::getTemperature(double* temperature) {
-  double out = analogRead(_pinNumber);
-  double resistance = RESISTANCE_0 * out / (4096.0 - out);
-  double temperatureK = (B * (TEMPERATURE_0 + K)) / (((TEMPERATURE_0 + K) * log(resistance / RESISTANCE_0)) + B);
+void Thermistor::getTemperature(float* temperature) {
+  float out = analogRead(_pinNumber);
+  float resistance = RESISTANCE_0 * out / (4096.0 - out);
+  float temperatureK = (B * (TEMPERATURE_0 + K)) / (((TEMPERATURE_0 + K) * log(resistance / RESISTANCE_0)) + B);
   *temperature = temperatureK - K;
 }
