+++ Projects/STM32duino/SensingModule/Thermistor.hpp
@@ -10,13 +10,13 @@ public:
   Thermistor(uint8_t pinNumber);
   void initialize();
 
-  void getTemperature(double* temperature);
+  void getTemperature(float* temperature);
 
 private:
-  const double B = 3380.0;
-  const double K = 273.15;
-  const double RESISTANCE_0 = 10000.0;
-  const double TEMPERATURE_0 = 25.0;
+  const float B = 3380.0;
+  const float K = 273.15;
+  const float RESISTANCE_0 = 10000.0;
+  const float TEMPERATURE_0 = 25.0;
 
   uint8_t _pinNumber;
 };
